TIDIJ-230 Quick fix
